+++ lua/limei/groups/treesitter.lua
@@ -6,6 +6,7 @@ function M.get(c)
     ["@variable"] = { link = "LimeiVariable" },
     ["@variable.builtin"] = { link = "@variable" },
     ["@variable.parameter"] = { link = "LimeiMuted" },
+    ["@variable.parameter.builtin"] = { link = "@variable.parameter" },
     ["@variable.member"] = { link = "@variable" },
     ["@property"] = { link = "@variable" },
     ["@label"] = { link = "LimeiMuted" },
@@ -14,28 +15,41 @@ function M.get(c)
     ["@function"] = { link = "LimeiFunction" },
     ["@function.builtin"] = { link = "@function" },
     ["@function.call"] = { link = "@function" },
+    ["@function.macro"] = { link = "@function" },
     ["@function.method"] = { link = "@function" },
     ["@function.method.call"] = { link = "@function" },
     ["@constructor"] = { link = "LimeiDeclaration" },
 
     -- Keywords and structure
     ["@attribute"] = { link = "LimeiDeclaration" },
+    ["@attribute.builtin"] = { link = "@attribute" },
     ["@keyword"] = { link = "LimeiKeyword" },
+    ["@keyword.coroutine"] = { link = "LimeiControl" },
     ["@keyword.function"] = { link = "LimeiDeclaration" },
     ["@keyword.modifier"] = { link = "LimeiMuted" },
     ["@keyword.operator"] = { link = "@keyword" },
+    ["@keyword.type"] = { link = "LimeiDeclaration" },
     ["@keyword.return"] = { link = "LimeiReturn" },
     ["@keyword.import"] = { link = "LimeiImport" },
     ["@keyword.conditional"] = { link = "LimeiControl" },
+    ["@keyword.conditional.ternary"] = { link = "@keyword.conditional" },
     ["@keyword.repeat"] = { link = "LimeiControl" },
+    ["@keyword.debug"] = { fg = c.warning },
     ["@keyword.exception"] = { link = "LimeiException" },
+    ["@keyword.directive"] = { link = "LimeiKeyword" },
+    ["@keyword.directive.define"] = { link = "LimeiDeclaration" },
 
     -- Strings and literals
     ["@string"] = { link = "LimeiString" },
     ["@string.documentation"] = { link = "LimeiDocComment" },
     ["@string.regexp"] = { link = "LimeiRegex" },
     ["@string.escape"] = { link = "LimeiEscape" },
+    ["@string.special"] = { link = "LimeiString" },
+    ["@string.special.symbol"] = { link = "LimeiConstant" },
+    ["@string.special.path"] = { link = "LimeiImport" },
+    ["@string.special.url"] = { fg = c.information, underline = true },
     ["@character"] = { link = "LimeiString" },
+    ["@character.special"] = { link = "LimeiEscape" },
 
     -- Numbers and logic
     ["@boolean"] = { link = "LimeiBoolean" },
@@ -51,6 +65,7 @@ function M.get(c)
     -- Constants and symbolic values
     ["@constant"] = { link = "LimeiConstant" },
     ["@constant.builtin"] = { link = "LimeiBuiltinConstant" },
+    ["@constant.macro"] = { link = "LimeiConstant" },
 
     -- Modules and namespaces
     ["@module"] = { link = "LimeiNamespace" },
@@ -58,6 +73,7 @@ function M.get(c)
 
     -- Operators and punctuation
     ["@operator"] = { link = "LimeiMuted" },
+    ["@punctuation"] = { link = "LimeiPunctuation" },
     ["@punctuation.delimiter"] = { link = "LimeiPunctuation" },
     ["@punctuation.bracket"] = { link = "LimeiMuted" },
     ["@punctuation.special"] = { link = "LimeiPunctuation" },
@@ -68,7 +84,7 @@ function M.get(c)
     ["@comment.error"] = { fg = c.error, bold = true },
     ["@comment.warning"] = { fg = c.warning, bold = true },
     ["@comment.todo"] = { fg = c.warning, bold = true },
-    ["@comment.note"] = { fg = c.literal, bold = true },
+    ["@comment.note"] = { fg = c.information, bold = true },
 
     -- Markup
     ["@markup.heading"] = { link = "Title" },
@@ -83,13 +99,32 @@ function M.get(c)
     ["@markup.raw"] = { link = "LimeiString" },
     ["@markup.list"] = { link = "LimeiPunctuation" },
     ["@markup.quote"] = { link = "LimeiMuted" },
+    ["@markup.math"] = { link = "LimeiNumber" },
     ["@markup.strong"] = { bold = true },
     ["@markup.italic"] = { italic = true },
+    ["@markup.strikethrough"] = { strikethrough = true },
+    ["@markup.underline"] = { underline = true },
+    ["@markup.link.label"] = { fg = c.information },
+    ["@markup.raw.block"] = { link = "@markup.raw" },
+    ["@markup.list.checked"] = { fg = c.success },
+    ["@markup.list.unchecked"] = { fg = c.warning },
 
     -- Diff
     ["@diff.plus"] = { link = "LimeiAdded" },
     ["@diff.minus"] = { link = "LimeiRemoved" },
     ["@diff.delta"] = { link = "LimeiChanged" },
+
+    -- Tags
+    ["@tag"] = { link = "LimeiDeclaration" },
+    ["@tag.builtin"] = { link = "@tag" },
+    ["@tag.attribute"] = { link = "LimeiVariable" },
+    ["@tag.delimiter"] = { link = "LimeiPunctuation" },
+
+    -- Query controls intentionally carry no visual styling.
+    ["@none"] = {},
+    ["@conceal"] = {},
+    ["@spell"] = {},
+    ["@nospell"] = {},
   }
 end